refactor feee and removed cruft
